Changed implementation of iterate Added nthiteration/3
I thought the implementation of `iterate/3` was slightly unintuitive.
`iterate(Fun, X, 10)` would apply the Fun at most *nine* times,
which would result in a resulting list of length 10.
Now, instead, it will apply the function at most 10 times which means
that the resulting list will be of length 11, which is also confusing
but slightly better in my opinion.
This means that `iterate(Fun, X, 0)` will now not return an empty list,
but a list with the original value.

The main reason for the change is because `nthiteration/3` was added,
which is equivalent to `lists:last(rlists:iterate(Fun, X, N))`,
in other words, the result of applying `Fun` N times.

Updated README.md to reflect these changes.